topology_coordinator: exec_global_command: make it optional to retake the guard
Currently, exec_global_command takes a group 0 guard, drops it and
retakes it after the command is finished. For current uses it is fine
from the correctness' point of view and, given that an operation can
take a long time, shorter duration of the guard improves the odds of the
operation succeeding.

However, this is not sufficient for cluster features because they will
need to execute a global barrier under the group 0 guard.

This commit modifies the interface of `exec_global_command` so that
dropping and retaking the guard is optional (the default is to retake
it).