raft: make a removed node a non-voter early
For removenode, we make a removed node a non-voter early. There is
no downside to it because the node is already dead. Moreover, it
improves availability in some situations. Consider a 4-node
cluster with one dead none. If we make the dead node a non-voter
at the beginning of removenode, group 0 will survive the death
of another node in the middle of removenode.