raft: make a decommissioning node a non-voter early
If we decommission a node when the number of nodes is even, we
make it a non-voter early to improve availability. All majorities
containing this node will remain majorities when we make this node
a non-voter and remove it from the set because the required size
of a majority decreases.