repair: generate random run_id across tables/ranges
Create a random sstables run_id for new table repairs
and reuse it as long as we detect disjoint token ranges.

If we detect in the range_tracker that a repair for
a new table has begun, or the repair_meta range overlaps
a previously repaired range in the table, reset the
tracked ranges and generate a new random run_id.

Note that this implementation presumes that repair
is scheduled one table at a time in each repair
master.  If multiple tables are to be repaired
in parallel, we'd need to track the ranges in each
table independently, but we don't have a good way
to determine when table-repair is done so to cleanup
this state.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>