topology_coordinator: account for features to enable in should_preempt_balancing
In #14722, a source of work was added to `handle_topology_transition`,
but the `should_preempt_balancing` function was not updated accordingly,
as is suggested by the comment in `handle_topology_transition`. This
omission happened due to a hasty rebase.

This commit fixes the issue, and now `should_preempt_balancing` will
return true if there are some features that should be enabled.