raft topology: support IPs for --ignore-dead-nodes
We want to stop supporting IPs for --ignore-dead-nodes in
raft_removenode and --ignore-dead-nodes-for-replace for
raft_replace. However, we shouldn't remove these features without
the deprecation period because the original removenode and
replace operations still support them. So, we add them for now.

Additionally, we modify test_raft_ignore_nodes.py so that it
verifies the added IP support.